+++ atelier/test.py
@@ -16,7 +16,7 @@ import glob
 from fnmatch import fnmatch
 import sys
 from setuptools import find_packages
-# from unipath import Path
+# from pathlib import Path
 import subprocess
 
 from atelier.utils import SubProcessParent